Demo code for the paper published on DSP
In the rar pakage, the is demo code for the paper "Variational Bayesian learning for removal of sparse impulsive noise from speech signals[J]. Digital Signal Processing, 2018, 73:106-116". The paper is focused on removal of impulsive noise from speech signals, and is related with the topic of speech enhancement. The speed of the code is slow, please be patient while running the code. If you use this code,please cite the paper.